+++ meteor-app/tests/cypress/integration/b2b/e2e-wwcc-2.js
@@ -17,6 +17,12 @@ Cypress.on('uncaught:exception', (err, runnable) => {
   // from a js library we are using
   return false
 })
+describe('Create an Cathrine King', function() {
+  it(`Creates an Cathrine`, function() {
+    cy.visit('/')
+    addCathrine()
+  })
+})
 
 // goes to the wwcc webpage
 describe('Visit WWCC webpage', function() {
@@ -29,7 +35,7 @@ describe('Visit WWCC webpage', function() {
       .click()
 
     // Adds a wwcc number that works
-    cy.get('button[about="Cathrine King-add"]')
+    cy.get('button[about="Cathrine Carmel-add"]')
       .contains('Add...')
       .should('be.enabled')
       .click()
@@ -49,7 +55,7 @@ describe('Visit WWCC webpage', function() {
 
     // Click the main button (not in the modal)
     //Clicks it again
-    cy.get('button[about="Cathrine King"]')
+    cy.get('button[about="Cathrine Carmel"]')
       .contains('Check')
       .should('be.enabled')
       .click()
@@ -60,10 +66,11 @@ describe('Visit WWCC webpage', function() {
     //  finds volunteer with wwcc number
     cy.get('[type="text"]')
       .clear()
-      .type('Cathrine King')
+      .type('Cathrine Carmel')
     cy.get('[type="text"]').clear()
     // At this point, only the card for our test subject should
     // be shown, and it should show the green tick to denote a valid WWC
     cy.get('[class="green check icon"]').should('exist')
+    rmCathrineKing()
   })
 })