+++ meteor-app/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added view of charges with 'refresh' button
 - Add reconciliation to charges and carts, improved layouts
 - Add admin button to edit user details
+- Changed input in add form to hid password
 - Add b2b member portal
 - Fix coding problems with certain test scripts
 - Add new service test scripts for input and UI
@@ -18,7 +19,7 @@
 - Fix the event import
 - Fix SMS sending to send message as typed
 - Fix SMS sending to detect price change and update DB to suit
-
+- Fix integration tests
 
 ## 1.2.4